Expo, code cleanup (please see main commit message for details)
(bzr r3700.1.1)

+++ plugins/expo/src/click_threshold/src/click-threshold.cpp
@@ -25,7 +25,7 @@
 #include "click-threshold.h"
 #include <stdlib.h>
 
-static const int DND_THRESHOLD = 5;
+static const unsigned short DND_THRESHOLD = 5;
 
 bool
 compiz::expo::clickMovementInThreshold(int previousX, int previousY,
